% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gap_plot}
\alias{gap_plot}
\title{Plot Gaps in a Time Series}
\usage{
gap_plot(
  station_no = "no_number",
  station_name = "no_name",
  time_vec,
  value,
  gap_info,
  interval_min
)
}
\arguments{
\item{value}{A numeric vector of corresponding measurements.}

\item{gap_info}{A data frame containing gap information. Must have at least
the columns \code{start} and \code{end} indicating the beginning and end of each gap.}

\item{time}{A vector of time points (e.g., POSIXct or numeric).}
}
\value{
Prints a ggplot2 object of the time series with gaps highlighted.
}
\description{
This function creates a plot of a time series with highlighted gaps.
Gaps are shown as red bars below the x-axis.
}
\details{
\itemize{
\item Points of the time series are plotted in blue.
\item The time series line is drawn in grey.
\item Gaps (from \code{gap_info}) are shown as red bars from y = 0 to y = -10.
\item Titles and axis labels are set automatically.
}
}
\examples{
time <- seq.POSIXt(from = as.POSIXct("2026-01-01 00:00"), 
                   to = as.POSIXct("2026-01-01 12:00"), by = "hour")
value <- sin(seq(0, 12, length.out = 13))
gap_info <- data.frame(start = as.POSIXct(c("2026-01-01 03:00", "2026-01-01 09:00")),
                       end   = as.POSIXct(c("2026-01-01 04:00", "2026-01-01 10:00")))
gap_plot(time, value, gap_info)

}
