% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aquazis_meta}
\alias{get_aquazis_meta}
\title{Get metadata (station information) from an AQUAZIS hub}
\usage{
get_aquazis_meta(hub, shared_data = NULL, logpath = NULL)
}
\arguments{
\item{hub}{Character. Base URL of the AQUAZIS hub (must not be \code{NULL} or empty).}

\item{shared_data}{Character (optional). Path to a directory where fallback metadata
is stored or should be saved as "aquazis_stations.rds".}

\item{logpath}{Character (optional). Path to a directory where warnings and errors
will be logged in "try.outFile".}
}
\value{
A tibble containing the metadata (\code{data}) from the AQUAZIS hub.
If the request fails and fallback data exists, the fallback is returned instead.
}
\description{
Retrieves metadata (e.g., station list) from a given AQUAZIS hub.
If the request fails, a fallback \code{.rds} file can be used (if \code{shared_data} is specified).
}
\examples{
\dontrun{
get_aquazis_meta("https://aquazis.example.com")
get_aquazis_meta("https://aquazis.example.com", shared_data = "data", logpath = "logs")
}

}
