% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aquazis_zr}
\alias{get_aquazis_zr}
\title{Get time series data (ZR) from an AQUAZIS hub}
\usage{
get_aquazis_zr(
  hub = NULL,
  zrid = NULL,
  begin = "",
  end = "",
  max_years = 1,
  pause_sec = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{hub}{Character. Base URL of the AQUAZIS hub.}

\item{zrid}{Character. ZR identifier (time series id) to request.}

\item{begin}{Character. Optional start date/time as string (default: "").}

\item{end}{Character. Optional end date/time as string (default: "").}
}
\value{
A JSON object (as a list) containing time series data.
}
\description{
Retrieves time series data for a given ZR-ID and optional time range.
}
\examples{
\dontrun{
get_aquazis_zr("https://aquazis.example.com",
               zrid = "12345",
               begin = "2020-01-01",
               end = "2020-12-31")
}

}
