% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_az_valid_to}
\alias{get_az_valid_to}
\title{Get Valid-To Timestamp for AQUAZIS Time Series}
\usage{
get_az_valid_to(
  hub,
  zrid,
  begin,
  end,
  intervall = "l",
  stepsize = 30,
  max_retries = 5
)
}
\arguments{
\item{hub}{Character. The hub identifier to query data from.}

\item{zrid}{Character or numeric. The ZR-ID used to identify the data series.}

\item{begin}{POSIXct. The start date/time of the data request.}

\item{end}{POSIXct. The end date/time of the data request.}

\item{intervall}{Character. The interval string for time series aggregation. Default is \code{"l"}.}

\item{stepsize}{Integer. Number of days to go further back in time when retrying. Default is \code{30}.}

\item{max_retries}{Integer. Maximum number of exponential backoff retries. Default is \code{5}.}
}
\value{
POSIXct. The timestamp indicating the "valid to" point, which is the last valid time before a missing data block.
}
\description{
This function retrieves and processes AQUAZIS time series data for a given hub and ZR-ID,
and determines the latest timestamp before a continuous block of missing data (\code{NA}s).
If the data is insufficient, it performs exponential backoff retries to fetch more historical data.
}
\examples{
\dontrun{
  hub <- "example_hub"
  zrid <- "123456"
  begin <- as.POSIXct("2025-01-01")
  end <- as.POSIXct("2025-08-01")
  valid_to <- get_az_valid_to(hub, zrid, begin, end)
  print(valid_to)
}

}
