% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_recent_eta_curves}
\alias{get_recent_eta_curves}
\title{Get most recent ETA curves}
\usage{
get_recent_eta_curves(vec)
}
\arguments{
\item{vec}{Character vector. Input vector with numeric values as strings.}
}
\value{
An integer vector with indices of the highest repeated two-digit value.
}
\description{
This function identifies two-digit values in a vector, finds those that occur
at least twice, and returns the indices of the highest such value.
}
\examples{
vec <- c("01", "02", "12", "12", "07", "12", "09")
get_recent_eta_curves(vec)
}
