% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aquazis_meta}
\alias{get_aquazis_meta}
\title{Get metadata from an AQUAZIS hub}
\usage{
get_aquazis_meta(hub)
}
\arguments{
\item{hub}{Character. Base URL of the AQUAZIS hub.}
}
\value{
A tibble containing metadata.
}
\description{
This function retrieves metadata from an AQUAZIS hub and returns it as a tibble.
}
\examples{
\dontrun{
get_aquazis_meta("https://aquazis.example.com")
}
}
