% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aquazis_zr}
\alias{get_aquazis_zr}
\title{Get time series data (ZR) from an AQUAZIS hub}
\usage{
get_aquazis_zr(hub, parameter)
}
\arguments{
\item{hub}{Character. Base URL of the AQUAZIS hub.}

\item{parameter}{List. Request parameters (e.g., \code{list(f_parameter="Abfluss")}).}
}
\value{
A JSON object (as a list) containing time series data.
}
\description{
This function retrieves time series data for a given parameter from the AQUAZIS hub.
}
\examples{
\dontrun{
get_aquazis_zr("https://aquazis.example.com", list(f_parameter="Abfluss"))
}
}
