% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aquazis_zrlist}
\alias{get_aquazis_zrlist}
\title{Get time series list (ZR) from an AQUAZIS hub}
\usage{
get_aquazis_zrlist(hub, st_id = "", parameter = "")
}
\arguments{
\item{hub}{Character. Base URL of the AQUAZIS hub.}

\item{parameter}{character. Request parameters (e.g., \code{c("Wasserstand)}).}
}
\description{
This function retrieves a list of time series from the AQUAZIS hub.
Optionally, a parameter can be provided to filter the request.
}
\examples{
\dontrun{
get_aquazis_zrlist("https://aquazis.example.com")
get_aquazis_zrlist("https://aquazis.example.com", list(f_parameter="Abflusskurve"))
}
}
